+++ game/content/ghplots/dd_intro.py
@@ -55,7 +55,7 @@ class DZDIntro_GetInTheMekShimli(Plot):
         myroom.contents.append(ghwaypoints.ClosedBoardingChute())
         myroom.contents.append(ghwaypoints.VentFan())
 
-        #myroom.contents.append(game.content.ghterrain.PersonalCargoContainerTerrset())
+        #myroom.contents.append(game.content.ghwaypoints.FusionCore())
 
         npc = self.register_element("SHERIFF",
                             gears.selector.random_character(45, local_tags=self.elements["LOCALE"].attributes,