+++ history.md
@@ -1,3 +1,5 @@
+* Fixed bug with text editor widget
+* Added more quests to Raid on Pirate's Point
 * Columns decor should not overwrite waypoint terrain
 * Fixed bug with non-singleton city tags in bugs
 * PC should now automatically reject most missions against favorable factions